Refactor nano evaluators to minimize main diff
