+++ sentence_transformers/evaluation/NanoBEIREvaluator.py
@@ -1,13 +1,24 @@
 from __future__ import annotations
 
+import logging
+import os
 from collections.abc import Callable
-from typing import Any, Literal
+from typing import TYPE_CHECKING, Any, Literal
 
 import numpy as np
 from torch import Tensor
+from tqdm import tqdm
 
-from sentence_transformers.evaluation.NanoEvaluator import NanoEvaluator
+from sentence_transformers import SentenceTransformer
+from sentence_transformers.evaluation.InformationRetrievalEvaluator import InformationRetrievalEvaluator
+from sentence_transformers.evaluation.SentenceEvaluator import SentenceEvaluator
 from sentence_transformers.similarity_functions import SimilarityFunction
+from sentence_transformers.util import is_datasets_available
+
+if TYPE_CHECKING:
+    from sentence_transformers.SentenceTransformer import SentenceTransformer
+
+logger = logging.getLogger(__name__)
 
 DatasetNameType = Literal[
     "climatefever",
@@ -25,7 +36,7 @@ DatasetNameType = Literal[
     "touche2020",
 ]
 
-DATASET_NAME_TO_HUMAN_READABLE: dict[str, str] = {
+DATASET_NAME_TO_HUMAN_READABLE = {
     "climatefever": "ClimateFEVER",
     "dbpedia": "DBPedia",
     "fever": "FEVER",
@@ -42,7 +53,7 @@ DATASET_NAME_TO_HUMAN_READABLE: dict[str, str] = {
 }
 
 
-class NanoBEIREvaluator(NanoEvaluator):
+class NanoBEIREvaluator(SentenceEvaluator):
     """
     This class evaluates the performance of a SentenceTransformer Model on the NanoBEIR collection of Information Retrieval datasets.
 
@@ -54,9 +65,6 @@ class NanoBEIREvaluator(NanoEvaluator):
     alongside many translated versions.
     This evaluator reports the same metrics as the :class:`~sentence_transformers.evaluation.InformationRetrievalEvaluator`
     (e.g., MRR, nDCG, Recall@k) for each dataset individually, as well as aggregated across all datasets.
-
-    This class preserves the NanoBEIR short-name API and delegates dataset-agnostic loading/aggregation mechanics to
-    :class:`~sentence_transformers.evaluation.NanoEvaluator`.
 
     Args:
         dataset_names (List[str]): The short names of the datasets to evaluate on (e.g., "climatefever", "msmarco").
@@ -200,6 +208,8 @@ class NanoBEIREvaluator(NanoEvaluator):
             pprint({key: value for key, value in results.items() if "ndcg@10" in key})
     """
 
+    information_retrieval_class = InformationRetrievalEvaluator
+
     def __init__(
         self,
         dataset_names: list[DatasetNameType | str] | None = None,
@@ -220,40 +230,274 @@ class NanoBEIREvaluator(NanoEvaluator):
         query_prompts: str | dict[str, str] | None = None,
         corpus_prompts: str | dict[str, str] | None = None,
         write_predictions: bool = False,
-    ) -> None:
+    ):
+        super().__init__()
         if dataset_names is None:
-            dataset_names = list(DATASET_NAME_TO_HUMAN_READABLE.keys())
+            dataset_names = self._get_default_dataset_names()
+        self.dataset_names = dataset_names
+        self.dataset_id = dataset_id
+        self.aggregate_fn = aggregate_fn
+        self.aggregate_key = aggregate_key
+        self.write_csv = write_csv
+        self.query_prompts = query_prompts
+        self.corpus_prompts = corpus_prompts
+        self.show_progress_bar = show_progress_bar
+        self.write_csv = write_csv
+        self.score_functions = score_functions
+        self.score_function_names = sorted(list(self.score_functions.keys())) if score_functions else []
+        self.main_score_function = main_score_function
+        self.truncate_dim = truncate_dim
+        self.name = f"{self._get_evaluator_name_root()}_{aggregate_key}"
+        if self.truncate_dim:
+            self.name += f"_{self.truncate_dim}"
 
-        super().__init__(
-            dataset_names=[str(name) for name in dataset_names],
-            dataset_id=dataset_id,
-            mrr_at_k=mrr_at_k,
-            ndcg_at_k=ndcg_at_k,
-            accuracy_at_k=accuracy_at_k,
-            precision_recall_at_k=precision_recall_at_k,
-            map_at_k=map_at_k,
-            show_progress_bar=show_progress_bar,
-            batch_size=batch_size,
-            write_csv=write_csv,
-            truncate_dim=truncate_dim,
-            score_functions=score_functions,
-            main_score_function=main_score_function,
-            aggregate_fn=aggregate_fn,
-            aggregate_key=aggregate_key,
-            query_prompts=query_prompts,
-            corpus_prompts=corpus_prompts,
-            write_predictions=write_predictions,
-            dataset_name_to_human_readable=DATASET_NAME_TO_HUMAN_READABLE,
-            split_prefix="Nano",
-            strict_dataset_name_validation=True,
-            auto_expand_splits_when_dataset_names_none=False,
-            name="NanoBEIR",
+        self.mrr_at_k = mrr_at_k
+        self.ndcg_at_k = ndcg_at_k
+        self.accuracy_at_k = accuracy_at_k
+        self.precision_recall_at_k = precision_recall_at_k
+        self.map_at_k = map_at_k
+
+        self._validate_dataset_names()
+        self._validate_prompts()
+        self._validate_mapping_splits()
+
+        ir_evaluator_kwargs = {
+            "mrr_at_k": mrr_at_k,
+            "ndcg_at_k": ndcg_at_k,
+            "accuracy_at_k": accuracy_at_k,
+            "precision_recall_at_k": precision_recall_at_k,
+            "map_at_k": map_at_k,
+            "show_progress_bar": show_progress_bar,
+            "batch_size": batch_size,
+            "write_csv": write_csv,
+            "truncate_dim": truncate_dim,
+            "score_functions": score_functions,
+            "main_score_function": main_score_function,
+            "write_predictions": write_predictions,
+        }
+        self.evaluators = [
+            self._load_dataset(name, **ir_evaluator_kwargs)
+            for name in tqdm(self.dataset_names, desc=self._get_loading_description(), leave=False)
+        ]
+
+        self.csv_file: str = f"{self._get_evaluator_name_root()}_evaluation_{aggregate_key}_results.csv"
+        self.csv_headers = ["epoch", "steps"]
+
+        self._append_csv_headers(self.score_function_names)
+
+    def _append_csv_headers(self, score_function_names):
+        for score_name in score_function_names:
+            for k in self.accuracy_at_k:
+                self.csv_headers.append(f"{score_name}-Accuracy@{k}")
+
+            for k in self.precision_recall_at_k:
+                self.csv_headers.append(f"{score_name}-Precision@{k}")
+                self.csv_headers.append(f"{score_name}-Recall@{k}")
+
+            for k in self.mrr_at_k:
+                self.csv_headers.append(f"{score_name}-MRR@{k}")
+
+            for k in self.ndcg_at_k:
+                self.csv_headers.append(f"{score_name}-NDCG@{k}")
+
+            for k in self.map_at_k:
+                self.csv_headers.append(f"{score_name}-MAP@{k}")
+
+    def __call__(
+        self,
+        model: SentenceTransformer,
+        output_path: str | None = None,
+        epoch: int = -1,
+        steps: int = -1,
+        *args,
+        **kwargs,
+    ) -> dict[str, float]:
+        per_metric_results = {}
+        per_dataset_results = {}
+        if epoch != -1:
+            if steps == -1:
+                out_txt = f" after epoch {epoch}"
+            else:
+                out_txt = f" in epoch {epoch} after {steps} steps"
+        else:
+            out_txt = ""
+        if self.truncate_dim is not None:
+            out_txt += f" (truncated to {self.truncate_dim})"
+        logger.info(
+            f"{self._get_evaluation_description()} Evaluation of the model on {self.dataset_names} dataset{out_txt}:"
         )
 
-    def _validate_dataset_names(self) -> None:
+        if self.score_functions is None:
+            self.score_functions = {model.similarity_fn_name: model.similarity}
+            self.score_function_names = [model.similarity_fn_name]
+            self._append_csv_headers(self.score_function_names)
+
+        num_underscores_in_name = self.name.count("_")
+        for evaluator in tqdm(self.evaluators, desc="Evaluating datasets", disable=not self.show_progress_bar):
+            logger.info(f"Evaluating {evaluator.name}")
+            evaluation = evaluator(model, output_path, epoch, steps)
+            for full_key, metric_value in evaluation.items():
+                metric = self._get_metric_from_full_key(evaluator.name, full_key, num_underscores_in_name)
+                if metric not in per_metric_results:
+                    per_metric_results[metric] = []
+                per_dataset_results[full_key] = metric_value
+                per_metric_results[metric].append(metric_value)
+
+        agg_results = {}
+        for metric in per_metric_results:
+            agg_results[metric] = self.aggregate_fn(per_metric_results[metric])
+
+        if output_path is not None and self.write_csv:
+            os.makedirs(output_path, exist_ok=True)
+            csv_path = os.path.join(output_path, self.csv_file)
+            if not os.path.isfile(csv_path):
+                fOut = open(csv_path, mode="w", encoding="utf-8")
+                fOut.write(",".join(self.csv_headers))
+                fOut.write("\n")
+
+            else:
+                fOut = open(csv_path, mode="a", encoding="utf-8")
+
+            output_data = [epoch, steps]
+            for name in self.score_function_names:
+                for k in self.accuracy_at_k:
+                    output_data.append(agg_results[f"{name}_accuracy@{k}"])
+
+                for k in self.precision_recall_at_k:
+                    output_data.append(agg_results[f"{name}_precision@{k}"])
+                    output_data.append(agg_results[f"{name}_recall@{k}"])
+
+                for k in self.mrr_at_k:
+                    output_data.append(agg_results[f"{name}_mrr@{k}"])
+
+                for k in self.ndcg_at_k:
+                    output_data.append(agg_results[f"{name}_ndcg@{k}"])
+
+                for k in self.map_at_k:
+                    output_data.append(agg_results[f"{name}_map@{k}"])
+
+            fOut.write(",".join(map(str, output_data)))
+            fOut.write("\n")
+            fOut.close()
+
+        if not self.primary_metric:
+            if self.main_score_function is None:
+                score_function = max(
+                    [(name, agg_results[f"{name}_ndcg@{max(self.ndcg_at_k)}"]) for name in self.score_function_names],
+                    key=lambda x: x[1],
+                )[0]
+                self.primary_metric = f"{score_function}_ndcg@{max(self.ndcg_at_k)}"
+            else:
+                self.primary_metric = f"{self.main_score_function.value}_ndcg@{max(self.ndcg_at_k)}"
+
+        avg_queries = np.mean([len(evaluator.queries) for evaluator in self.evaluators])
+        avg_corpus = np.mean([len(evaluator.corpus) for evaluator in self.evaluators])
+        logger.info(f"Average Queries: {avg_queries}")
+        logger.info(f"Average Corpus: {avg_corpus}\n")
+
+        for name in self.score_function_names:
+            logger.info(f"Aggregated for Score Function: {name}")
+            for k in self.accuracy_at_k:
+                logger.info("Accuracy@{}: {:.2f}%".format(k, agg_results[f"{name}_accuracy@{k}"] * 100))
+
+            for k in self.precision_recall_at_k:
+                logger.info("Precision@{}: {:.2f}%".format(k, agg_results[f"{name}_precision@{k}"] * 100))
+                logger.info("Recall@{}: {:.2f}%".format(k, agg_results[f"{name}_recall@{k}"] * 100))
+
+            for k in self.mrr_at_k:
+                logger.info("MRR@{}: {:.4f}".format(k, agg_results[f"{name}_mrr@{k}"]))
+
+            for k in self.ndcg_at_k:
+                logger.info("NDCG@{}: {:.4f}".format(k, agg_results[f"{name}_ndcg@{k}"]))
+
+            for k in self.map_at_k:
+                logger.info("MAP@{}: {:.4f}".format(k, agg_results[f"{name}_map@{k}"]))
+
+        agg_results = self.prefix_name_to_metrics(agg_results, self.name)
+        self.store_metrics_in_model_card_data(model, agg_results, epoch, steps)
+
+        per_dataset_results.update(agg_results)
+
+        return per_dataset_results
+
+    def _get_human_readable_name(self, dataset_name: DatasetNameType | str) -> str:
+        human_readable_name = self._get_split_name(dataset_name)
+
+        if self.truncate_dim is not None:
+            human_readable_name += f"_{self.truncate_dim}"
+        return human_readable_name
+
+    def _load_dataset(
+        self, dataset_name: DatasetNameType | str, **ir_evaluator_kwargs
+    ) -> InformationRetrievalEvaluator:
+        split_name = self._get_split_name(dataset_name)
+
+        corpus = self._load_dataset_subset_split(
+            self._get_corpus_subset_name(),
+            split=split_name,
+            required_columns=["_id", "text"],
+        )
+        queries = self._load_dataset_subset_split(
+            self._get_queries_subset_name(),
+            split=split_name,
+            required_columns=["_id", "text"],
+        )
+        qrels = self._load_dataset_subset_split(
+            self._get_qrels_subset_name(),
+            split=split_name,
+            required_columns=["query-id", "corpus-id"],
+        )
+
+        corpus_dict = {sample["_id"]: sample["text"] for sample in corpus if len(sample["text"]) > 0}
+        queries_dict = {sample["_id"]: sample["text"] for sample in queries if len(sample["text"]) > 0}
+
+        qrels_dict = {}
+        for sample in qrels:
+            corpus_ids = sample.get("corpus-id")
+            if sample["query-id"] not in qrels_dict:
+                qrels_dict[sample["query-id"]] = set()
+
+            if isinstance(corpus_ids, list):
+                qrels_dict[sample["query-id"]].update(corpus_ids)
+            else:
+                qrels_dict[sample["query-id"]].add(corpus_ids)
+
+        if self.query_prompts is not None:
+            ir_evaluator_kwargs["query_prompt"] = self._get_prompt_for_dataset(self.query_prompts, dataset_name)
+        if self.corpus_prompts is not None:
+            ir_evaluator_kwargs["corpus_prompt"] = self._get_prompt_for_dataset(self.corpus_prompts, dataset_name)
+        human_readable_name = self._get_human_readable_name(dataset_name)
+        return self.information_retrieval_class(
+            queries=queries_dict,
+            corpus=corpus_dict,
+            relevant_docs=qrels_dict,
+            name=human_readable_name,
+            **ir_evaluator_kwargs,
+        )
+
+    def _load_dataset_subset_split(self, subset: str, split: str, required_columns: list[str]):
+        if not is_datasets_available():
+            raise ValueError(
+                "datasets is not available. Please install it to use the NanoBEIREvaluator via `pip install datasets`."
+            )
+        from datasets import load_dataset
+
+        try:
+            dataset = load_dataset(self.dataset_id, subset, split=split)
+        except Exception as e:
+            raise ValueError(
+                f"Could not load subset '{subset}' split '{split}' from dataset '{self.dataset_id}'."
+            ) from e
+
+        if missing_columns := set(required_columns) - set(dataset.column_names):
+            raise ValueError(
+                f"Subset '{subset}' split '{split}' from dataset '{self.dataset_id}' is missing required columns: {list(missing_columns)}."
+            )
+        return dataset
+
+    def _validate_dataset_names(self):
         if len(self.dataset_names) == 0:
             raise ValueError("dataset_names cannot be empty. Use None to evaluate on all datasets.")
-
         missing_datasets = [
             dataset_name
             for dataset_name in self.dataset_names
@@ -265,14 +509,79 @@ class NanoBEIREvaluator(NanoEvaluator):
                 f"Valid dataset names are: {list(DATASET_NAME_TO_HUMAN_READABLE.keys())}"
             )
 
+    def _validate_prompts(self):
+        error_msg = ""
+        if self.query_prompts is not None:
+            if isinstance(self.query_prompts, str):
+                self.query_prompts = {dataset_name: self.query_prompts for dataset_name in self.dataset_names}
+            elif missing_query_prompts := [
+                dataset_name
+                for dataset_name in self.dataset_names
+                if self._get_prompt_for_dataset(self.query_prompts, dataset_name) is None
+            ]:
+                error_msg += f"The following datasets are missing query prompts: {missing_query_prompts}\n"
+
+        if self.corpus_prompts is not None:
+            if isinstance(self.corpus_prompts, str):
+                self.corpus_prompts = {dataset_name: self.corpus_prompts for dataset_name in self.dataset_names}
+            elif missing_corpus_prompts := [
+                dataset_name
+                for dataset_name in self.dataset_names
+                if self._get_prompt_for_dataset(self.corpus_prompts, dataset_name) is None
+            ]:
+                error_msg += f"The following datasets are missing corpus prompts: {missing_corpus_prompts}\n"
+
+        if error_msg:
+            raise ValueError(error_msg.strip())
+
+    def _get_default_dataset_names(self) -> list[str]:
+        return list(DATASET_NAME_TO_HUMAN_READABLE.keys())
+
+    def _get_evaluator_name_root(self) -> str:
+        return "NanoBEIR"
+
+    def _get_evaluation_description(self) -> str:
+        return "NanoBEIR"
+
+    def _get_loading_description(self) -> str:
+        return "Loading NanoBEIR datasets"
+
+    def _get_split_name(self, dataset_name: DatasetNameType | str) -> str:
+        return f"Nano{DATASET_NAME_TO_HUMAN_READABLE[dataset_name.lower()]}"
+
+    def _get_corpus_subset_name(self) -> str:
+        return "corpus"
+
+    def _get_queries_subset_name(self) -> str:
+        return "queries"
+
+    def _get_qrels_subset_name(self) -> str:
+        return "qrels"
+
+    def _get_prompt_for_dataset(
+        self, prompt_mapping: dict[str, str], dataset_name: DatasetNameType | str
+    ) -> str | None:
+        return prompt_mapping.get(dataset_name, None)
+
+    def _get_metric_from_full_key(self, evaluator_name: str, full_key: str, num_underscores_in_name: int) -> str:
+        del evaluator_name
+        splits = full_key.split("_", maxsplit=num_underscores_in_name)
+        return splits[-1]
+
+    def _validate_mapping_splits(self) -> None:
+        pass
+
+    def store_metrics_in_model_card_data(self, *args, **kwargs):
+        # Only store metrics in the model card data if there is more than one dataset.
+        # Otherwise the e.g. mean scores for NanoBEIR are the same as the scores for
+        # the single dataset, and we'd end up with duplicate entries.
+        if len(self.dataset_names) > 1:
+            super().store_metrics_in_model_card_data(*args, **kwargs)
+
     def get_config_dict(self) -> dict[str, Any]:
-        config_dict: dict[str, Any] = {
-            "dataset_names": self.dataset_names,
-            "dataset_id": self.dataset_id,
-        }
+        config_dict = {"dataset_names": self.dataset_names, "dataset_id": self.dataset_id}
         config_dict_candidate_keys = ["truncate_dim", "query_prompts", "corpus_prompts"]
         for key in config_dict_candidate_keys:
-            value = getattr(self, key)
-            if value is not None:
-                config_dict[key] = value
+            if getattr(self, key) is not None:
+                config_dict[key] = getattr(self, key)
         return config_dict
